feat(core): new `utils` mod
