+++ api-server/llama-core/src/lib.rs
@@ -4,6 +4,7 @@ pub mod embeddings;
 pub mod error;
 pub mod models;
 pub mod rag;
+pub(crate) mod utils;
 
 pub use error::LlamaCoreError;
 
@@ -29,6 +30,9 @@ pub struct Metadata {
     // * Plugin parameters (used by this plugin):
     #[serde(rename = "enable-log")]
     pub log_enable: bool,
+    // this field not defined for the beckend plugin
+    #[serde(skip_serializing)]
+    pub log_prompts: bool,
     // #[serde(rename = "enable-debug-log")]
     // pub debug_log: bool,
     // #[serde(rename = "stream-stdout")]
@@ -149,34 +153,6 @@ impl Graph {
     pub fn finish_single(&mut self) -> Result<(), WasiNnError> {
         self.context.fini_single()
     }
-}
-
-pub(crate) fn print_log_begin_separator(
-    title: impl AsRef<str>,
-    ch: Option<&str>,
-    len: Option<usize>,
-) -> usize {
-    let title = format!(" [LOG: {}] ", title.as_ref());
-
-    let total_len: usize = len.unwrap_or(100);
-    let separator_len: usize = (total_len - title.len()) / 2;
-
-    let ch = ch.unwrap_or("-");
-    let mut separator = "\n\n".to_string();
-    separator.push_str(ch.repeat(separator_len).as_str());
-    separator.push_str(&title);
-    separator.push_str(ch.repeat(separator_len).as_str());
-    separator.push('\n');
-    println!("{}", separator);
-    total_len
-}
-
-pub(crate) fn print_log_end_separator(ch: Option<&str>, len: Option<usize>) {
-    let ch = ch.unwrap_or("-");
-    let mut separator = "\n\n".to_string();
-    separator.push_str(ch.repeat(len.unwrap_or(100)).as_str());
-    separator.push('\n');
-    println!("{}", separator);
 }
 
 /// Initialize the core context
